+++ public/templates/list-types/grid.php
@@ -370,7 +370,7 @@ $view_type_class = ' view-type-' . str_replace('_', '', $view_type);
                             <?php endif; ?>
                         <?php endif; ?>
                         <?php if (!$is_location_taxonomy && $show_location_room) : ?>
-                            <?php if ($show_location_name || $show_location_freetext) : ?> — <?php endif; ?>
+                            <?php if ($show_location_name || $show_location_freetext) : ?> - <?php endif; ?>
                             <?php echo esc_html($location_room); ?>
                         <?php endif; ?>
                     </strong>